Rename RandomizeEnv.py to randomized_env.py
All the python files have file names in lowercase. To keep this
standard, refactor RandomizeEnv.py to randomized_env.py.
